pyproject.toml
[project]
name = "ebay-seller-mcp"
version = "0.2.0"
description = "MCP server for eBay seller listing management. Built for Claude Code."
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.0",
    "ebaysdk>=2.2.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.18.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
# Exclude vendored scripts (maintained in dotfiles, drift-checked here).
# Modifying them locally would break the secret-rules-drift pre-commit hook.
extend-exclude = [
    "scripts/check-public-repo-secrets.py",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-x --tb=line -q"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyyaml>=6.0.3",
    "ruff>=0.15.10",
    "pytest>=8.0.0",
]
